add proper init/close functions
This way, the minor mode has no effect when it's not active.  Not sure
that's really what you want?  After all, some of these hooks have
benefits even when the timer is not actively engaged.  I guess the
best solution would be to have two minors mode, one of which requires
the other?  The first implements the general improvements, the second
activates the timer?  Think about it some more.