add support for buffer names that contain tildes
This is useful for using preview with buffers resulting from, e.g.,
vc-version-ediff.  The issue is that the tildes are escape characters
in tex, so the !name(...) commands in _region_.tex get muddled in
_region_.log.  The implementation consists of removing the tildes but
storing the file name and corresponding buffer in a dictionary.  The
dictionary is never freed, but only contains one entry per buffer that
gets previewed, so maybe it's OK.  Would be better to have it cleaned
up now and then.